app/requirements.txt
anyio==3.7.0
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.1.0
cycler==0.11.0
exceptiongroup==1.1.1
fonttools==4.40.0
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
kiwisolver==1.4.4
numpy==1.25.0
packaging==23.1
Pillow==9.5.0
pyparsing==3.1.0
pyTelegramBotAPI==4.12.0
python-dateutil==2.8.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
urllib3==2.0.3
Pillow